basic03: change map to be an array per xdp_action
This is to keep it compatible with future planned changes.
Where the xdp_stats program need to be able to get stats
from the packet0X lessons.

We could make it an assignment to extract other keys that XDP_PASS,
which is the simple approach in xdp_load_and_stats.c for now.

Signed-off-by: Jesper Dangaard Brouer <brouer@redhat.com>