Improve on Makefile common.mk
As discussed with Toke other parts of common.mk depend on $(COMMON_DIR),
not only $(COMMON_MK), as done in commit 20dae1dcaae5 with Makefile ifdef.

It is much simpler to use =? and define COMMON_DIR if it was not
defined by including Makefile.  This also takes care of other
parts that depend on COMMON_DIR beind defined.

Fixes: 20dae1dcaae5 ("Add build dependency if common.mk change")
Suggested-by: Toke Høiland-Jørgensen <toke@redhat.com>
Signed-off-by: Jesper Dangaard Brouer <brouer@redhat.com>