setup_dependencies: add link to Netronome's bpftool .deb package
bpftool is not packaged for Debian/Ubuntu at this time. However,
Netronome provides a .deb package for it (statically linked executable),
which might help people willing to try the tool without downloading the
kernel tree to build it. Let's add the link to that package in the
documentation (this should be removed when bpftool gets proper packaging
in Debian).