packet01: use struct for next header pointer iterator
This is basically C syntax-sugar for the samething that is
currently implemented, there should be no functional change.

This is an attempt to make the code easier to understand for
people that are not C-experts. While the code is correct, I fear
that people will be scared when seeing a void **ptr.

When introducing a simple struct with a pointer, we can make it
more clear that we are looking at the pointer value in the bound
check

Signed-off-by: Jesper Dangaard Brouer <brouer@redhat.com>