testenv: fix the --legacy-ip option implementation
The implementation of the '--legacy-ip' option was broken in two places.

First, the 'ip route' program was called to add a IPv4 route with a network
prefix ending with the IPv6 '::' notation.

Second, the 'get_num' function returns the number of a test environment in hex,
and this return value is used when defining IPv6 and/or IPv4 network prefixes.
In the IPv4 case we need to convert it to decimal, or some prefixes, starting
from the tenth test environment, e.g., '10.11.a' or '10.11.b', will be incorrect.

Signed-off-by: Anton Protopopov <a.s.protopopov@gmail.com>