packet-solutions: extract solutions to separate files
In order to split solutions for different lessons we move XDP code to
individual files.  This commit creates files containing solutions for the
packet03-redirecting and packet02-rewriting lessons.

In particular, add a solution for the first assignment of the packet-02 lesson:
rewrite UDP and TCP ports for incoming packets.

The packet01-parsing is pretty simple and examples of using parsing functions
can be found in all consequent examples.

Signed-off-by: Anton Protopopov <a.s.protopopov@gmail.com>