Create parse_ethhdr_vlan() in parsing_helpers.h
This is based on same function from packet-solutions/xdp_vlan02_kern.c
that also extracts VLANs.

It needs to be removed from packet-solutions/xdp_vlan02_kern.c, but
for now its just ifdef defined out-of-code.  Need to run some tests
and look at BPF byte-code to make sure it does the right thing.

The original parse_ethhdr() is implemented by calling parse_ethhdr_vlan()
with NULL argument for stucture to collect IDs into.  The compiler
should remove the parts that are not needed.

Signed-off-by: Jesper Dangaard Brouer <brouer@redhat.com>