helpers: Update license to dual-licensed GPL and BSD
We want to encourage reuse of the helper functions, so add a BSD licensing
option as well. Also, update the GPL specification to 2.0-or-later, since
the GPL-2.0 SPDX tag is deprecated, and with the BSD option, licensing as
2.0-only doesn't really make sense, since the BSD license permits inclusion
in GPL-3 works anyway.

All copyright holders of the affected files have signed off on this change,
as per the discussion in #80.

Fixes #80.

Signed-off-by: Toke Høiland-Jørgensen <toke@redhat.com>