testenv: Set ulimit to unlimited instead of a fixed number
Raising the ulimit to a fixed value doesn't work reliably: When we're only
loading a single program it works fine, but the limit is global for the
system, so if something else (e.g., another map or program) is already
loaded, the ulimit requirement is higher.

Rather than trying to guess, just set the memlock ulimit to unlimited.

Fixes #86.

Signed-off-by: Toke Høiland-Jørgensen <toke@redhat.com>