experiment01-tailgrow: try recommentation but it fails
This shows the code after Daniel's idea for change in PR #123

The code can now pass the verifier checks, but it will drop
all packets, because the offset is the packet lenght and
doing +1 will put us over the packet lenght.  Thus, the
if-check against data_end will cause a XDP_DROP.

The +1 is needed to pass the verifier checks.

Using XDP_DROP for this case, to make it easier to identify
the problem via using xdpdump[1]

 ./xdpdump -v -i mlx5p1 --rx-capture entry,exit -x

[1] https://github.com/xdp-project/xdp-tools/tree/master/xdp-dump

Signed-off-by: Jesper Dangaard Brouer <brouer@redhat.com>