common.mk: Make sure $OBJDIR is not overridden for libbpf compile
If $OBJDIR is set in the environment, that will be pulled in by the libbpf
Makefile, which will result in objects ending up in the wrong directory
where the tutorial bits can't find them. Make sure this doesn't happen by
explicitly setting OBJDIR=. as a 'make' parameter when building libbpf.

Reported-by: Steve Dickson <steved@redhat.com>
Signed-off-by: Toke Høiland-Jørgensen <toke@redhat.com>