testenv: rework veth creating and config
Instead of creating the veth in the root namespace and then moving and
renaming it, just create it directly in the target namespace with the right
name from the beginning. This requires a bit of rework of the settings
afterwards, but should hopefully fix any inconsistency issues such as the
MAC address issue reported in #159.

Signed-off-by: Toke Høiland-Jørgensen <toke@redhat.com>