testenv: wait for interface addresses to settle during setup
When configuring IPv6 addresses on an interface, they can be marked as
"tentative" until duplicate address detection (DAD) settles, which means
ping won't work. We try to turn off DAD, but that doesn't always work
either, depending on the global settings on the host; so add an explicit
wait after interface config.

Fixes #159.

Signed-off-by: Toke Høiland-Jørgensen <toke@redhat.com>