packet03: Use subnet numbers 1 and 2 in example command and output
The subnet numbers used by the testenv script when setting up the
environment can vary. The example commands in the packet03 assignment used
subnet numbers 10 and 11, but since that only matches if a lot of
namespaces are setup at once, change it to 1 and 2 to match the most common
setup where only two namespaces are created.

Fixes #221.

Signed-off-by: Toke Høiland-Jørgensen <toke@redhat.com>