tracing04-xdp-tcpdump: update value type to struct S
Commit 50703ef56dc3 ("tracing04: update for libbpf 1.0+ support) updated
the bpf map from the legacy to the BTF format. Unfortunately, the
value_size field was not updated to value type. The code works because
struct S has the same size as __u32 (and we can freely typecast with C).

Therefore, update the code so that the value type match the expected type and
match the text to reference the value type (BTF format), instead of
value size (legacy format).

Signed-off-by: Rodrigo Cataldo <rodrigo.cadore@l-acoustics.com>